[HBASE-5081] Distributed log splitting deleteNode races against splitLog retry
Summary:
fixes the race that occurs when a failed task is retried and the older
incarnation of the failed task hasn't been properly cleaned up.

removes OrphanLogException that was being thrown. This is no longer needed
after HBASE-2... (rename logdir to logdir-splitting)

introduces a new behavior - what if the zk task node is manually deleted?
SplitLogManager will assume that the task is done. (It will not remove or move
any log files)

Test Plan:
tested the changed tests TestDistributedLogSplitting and TestSplitLogManager

this has been cluster tested in HBASE-5081 on open source trunk

Reviewers: liyintang, kannan, kranganathan

Reviewed By: kannan

CC: hbase-eng@lists

Differential Revision: https://phabricator.fb.com/D385894

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1239785 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/regionserver/wal/HLog.java
@@ -1831,7 +1831,7 @@ public class HLog implements Syncable {
   static void archiveLogs(final List<Path> corruptedLogs,
       final List<Path> processedLogs, final Path oldLogDir,
       final FileSystem fs, final Configuration conf)
-  throws IOException{
+  throws IOException {
     final Path corruptDir = new Path(conf.get(HConstants.HBASE_DIR),
       conf.get("hbase.regionserver.hlog.splitlog.corrupt.dir", ".corrupt"));
 
@@ -1841,21 +1841,27 @@ public class HLog implements Syncable {
     if (!fs.exists(oldLogDir) && !fs.mkdirs(oldLogDir)) {
       LOG.warn("Unable to mkdir " + oldLogDir);
     }
+    // this method can get restarted or called multiple times for archiving
+    // the same log files.
     for (Path corrupted: corruptedLogs) {
       Path p = new Path(corruptDir, corrupted.getName());
-      if (!fs.rename(corrupted, p)) {
-        LOG.warn("Unable to move corrupted log " + corrupted + " to " + p);
-      } else {
-        LOG.info("Moving corrupted log " + corrupted + " to " + p);
+      if (fs.exists(corrupted)) {
+        if (!fs.rename(corrupted, p)) {
+          LOG.warn("Unable to move corrupted log " + corrupted + " to " + p);
+        } else {
+          LOG.warn("Moving corrupted log " + corrupted + " to " + p);
+        }
       }
     }
 
     for (Path p: processedLogs) {
       Path newPath = getHLogArchivePath(oldLogDir, p);
-      if (!fs.rename(p, newPath)) {
-        LOG.warn("Unable to move processed log " + p + " to " + newPath);
-      } else {
-        LOG.info("Archived processed log " + p + " to " + newPath);
+      if (fs.exists(p)) {
+        if (!fs.rename(p, newPath)) {
+          LOG.warn("Unable to move  " + p + " to " + newPath);
+        } else {
+          LOG.debug("Archived processed log " + p + " to " + newPath);
+        }
       }
     }
   }
